+++ src/OpenIdConnect.php
@@ -260,7 +260,7 @@ class OpenIdConnect extends OAuth2
     {
         return $this->api($this->getConfigParam('userinfo_endpoint'), 'GET');
     }
-	
+
     /**
      * {@inheritdoc}
      */
@@ -317,25 +317,6 @@ class OpenIdConnect extends OAuth2
     /**
      * {@inheritdoc}
      */
-    protected function defaultReturnUrl()
-    {
-        $params = Yii::$app->getRequest()->getQueryParams();
-        // OAuth2 specifics :
-        unset($params['code']);
-        unset($params['state']);
-        // OpenIdConnect specifics :
-        unset($params['nonce']);
-        unset($params['authuser']);
-        unset($params['session_state']);
-        unset($params['prompt']);
-        $params[0] = Yii::$app->controller->getRoute();
-
-        return Yii::$app->getUrlManager()->createAbsoluteUrl($params);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     protected function createToken(array $tokenConfig = [])
     {
         if ($this->validateJws) {
@@ -378,8 +359,9 @@ class OpenIdConnect extends OAuth2
      * Validates the claims data received from OpenID provider.
      * @param array $claims claims data.
      * @throws HttpException on invalid claims.
+     * @since 2.2.3
      */
-    private function validateClaims(array $claims)
+    protected function validateClaims(array $claims)
     {
         if (!isset($claims['iss']) || (strcmp(rtrim($claims['iss'], '/'), rtrim($this->issuerUrl, '/')) !== 0)) {
             throw new HttpException(400, 'Invalid "iss"');