+++ src/clients/VKontakte.php
@@ -76,7 +76,7 @@ class VKontakte extends OAuth2
      * @see https://vk.com/dev/versions
      * @since 2.1.4
      */
-    public $apiVersion = '5.0';
+    public $apiVersion = '5.95';
 
 
     /**